requirements.txt
selenium
pytest
pytest-html
pytest-xdist
behave
python-dotenv
webdriver-manager